+++ Runtime/Core/ProBuilderMesh.cs
@@ -172,28 +172,39 @@ namespace UnityEngine.ProBuilder
         }
 #pragma warning restore 109
 
+        internal const ushort k_UnitializedVersionIndex = 0;
+
         /// <summary>
         /// Tracks each time the ToMesh() and Refresh() functions are called to modify the mesh.
         /// This is a simple uint number used to check whether two versions of the ProBuilderMesh are the same or not.
         /// </summary>
         [SerializeField]
-        ushort m_VersionIndex = 0;
+        ushort m_VersionIndex = k_UnitializedVersionIndex;
+
+        // k_NewMeshVersionIndex is a reserved value indicating that this is a new mesh instance. It is used to
+        // differentiate between mesh creation and duplication.
+        [NonSerialized]
+        ushort m_InstanceVersionIndex = k_UnitializedVersionIndex;
+
         internal ushort versionIndex => m_VersionIndex;
+        internal ushort nonSerializedVersionIndex => m_InstanceVersionIndex;
 
         internal struct NonVersionedEditScope : IDisposable
         {
-            ProBuilderMesh m_Mesh;
-            ushort m_VersionIndex;
+            readonly ProBuilderMesh m_Mesh;
+            readonly ushort m_VersionIndex, m_InstanceVersionIndex;
 
             public NonVersionedEditScope(ProBuilderMesh mesh)
             {
                 m_Mesh = mesh;
                 m_VersionIndex = mesh.versionIndex;
+                m_InstanceVersionIndex = mesh.m_InstanceVersionIndex;
             }
 
             public void Dispose()
             {
                 m_Mesh.m_VersionIndex = m_VersionIndex;
+                m_Mesh.m_InstanceVersionIndex = m_InstanceVersionIndex;
             }
         }
 
@@ -964,7 +975,7 @@ namespace UnityEngine.ProBuilder
             }
 
             set
-            { 
+            {
                 m_Mesh = value;
 #if UNITY_EDITOR
                 UnityEditor.EditorUtility.SetDirty(this);
@@ -973,10 +984,8 @@ namespace UnityEngine.ProBuilder
             }
         }
 
-        internal int id
-        {
-            get { return gameObject.GetInstanceID(); }
-        }
+        [Obsolete("InstanceID is not used to track mesh references as of 2023/04/12")]
+        internal int id => gameObject.GetInstanceID();
 
         /// <summary>
         /// Gets a flag that indicates whether the <see cref="UnityEngine.Mesh" /> is in sync with the ProBuilderMesh.
@@ -992,12 +1001,13 @@ namespace UnityEngine.ProBuilder
                 if (mesh == null)
                     return MeshSyncState.Null;
 
-                int meshNo;
-
-                int.TryParse(mesh.name.Replace("pb_Mesh", ""), out meshNo);
-
-                if (meshNo != id)
-                    return MeshSyncState.InstanceIDMismatch;
+                // if the local version index is uninitialized, that means no edits have occurred since loading this
+                // mesh. it could mean a new mesh was created and ToMesh has not been called yet, or that this mesh was
+                // copy/pasted from an existing instance.
+                // in the latter case, it is handled by listening to the ObjectChangeEvent stream in HierarchyListener.
+                // this function only cares about the case where modifications have been made during the current session.
+                if (m_VersionIndex != m_InstanceVersionIndex && m_InstanceVersionIndex != k_UnitializedVersionIndex)
+                    return MeshSyncState.NeedsRebuild;
 
                 return mesh.uv2 == null ? MeshSyncState.Lightmap : MeshSyncState.InSync;
             }